[Bronze V] Title: Multiply, Time: 200 ms, Memory: 17692 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.java
@@ -1,13 +1,12 @@
 import java.util.Scanner;
-
 public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		int a = sc.nextInt();
 		int b = sc.nextInt();		
-		int N = sc.nextInt();
-		int M = sc.nextInt();
-		int result = N*M;
+		long N = sc.nextInt();
+		long M = sc.nextInt();
+		long result = N*M;
 		System.out.println(result);
 	}
 }
